requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==21.4.0
    # via pytest
django==3.2.14
    # via
    #   -r requirements.in
    #   djangohexadecimalfield
djangohexadecimalfield==0.0.3
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pillow==9.2.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via django
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via pytest
